docker/ga3-microservices/service/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.1
tinydb==4.8.0
typing_extensions==4.12.2
